+++ errors.go
@@ -84,6 +84,8 @@ var (
 
 	ErrInvalidProjectID = errors.New("invalid value for project ID")
 
+	ErrInvalidPagination = errors.New("invalid value for page size or number")
+
 	ErrInvalidRunTaskCategory = errors.New(`category must be "task"`)
 
 	ErrInvalidRunTaskID = errors.New("invalid value for run task ID")