+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+## Enhancements
+
+* Add `Projects` interface with `list`, `read`, `update`, `delete` methods and integration tests by @hs26gill and @mwudka [#564](https://github.com/hashicorp/go-tfe/pull/564)
+
 # v1.13.0
 
 ## Bug Fixes